Move determine_unit_from_results
